+++ packages/lumx-vue/src/components/progress/index.ts
@@ -1,0 +1,7 @@
+export { default as ProgressCircular } from './ProgressCircular';
+export type { ProgressCircularProps, ProgressCircularSize } from './ProgressCircular';
+export { CLASSNAME as PROGRESS_CIRCULAR_CLASSNAME, COMPONENT_NAME as PROGRESS_CIRCULAR_COMPONENT_NAME } from './ProgressCircular';
+
+export { default as ProgressLinear } from './ProgressLinear';
+export type { ProgressLinearProps } from './ProgressLinear';
+export { CLASSNAME as PROGRESS_LINEAR_CLASSNAME, COMPONENT_NAME as PROGRESS_LINEAR_COMPONENT_NAME } from './ProgressLinear';