+++ packages/lumx-vue/src/index.ts
@@ -14,6 +14,7 @@ export * from './components/input-helper';
 export * from './components/input-label';
 export * from './components/link';
 export * from './components/message';
+export * from './components/progress';
 export * from './components/radio-button';
 export * from './components/skeleton';
 export * from './components/switch';