+++ packages/site-demo/content/product/components/progress/vue/demo2.vue
@@ -1,0 +1,9 @@
+<template>
+    <ProgressLinear :theme="theme" />
+</template>
+
+<script setup lang="ts">
+import { ProgressLinear, type Theme } from '@lumx/vue';
+
+defineProps<{ theme?: Theme }>();
+</script>